all tests passing and refactored from container to components
