+++ src/components/Header/index.scss
@@ -40,7 +40,6 @@
 
   &__name {
     font-family: $alt-font-family;
-    font-weight: bold;
     font-size: ($base-font-size * 1.25);
     max-width: 100px;
