+++ src/components/ProjectDashboardInfo/index.scss
@@ -14,6 +14,8 @@
   }
 
   &__title {
+    line-height: ($base-line-height - .2);
+    font-weight: 500;
     font-size: 36px;
 
     @media screen and (min-width: 768px) {
@@ -22,6 +24,7 @@
   }
 
   &__description {
+    font-weight: 300;
     margin-top: ($base-font-size * .5);
 
     @media screen and (min-width: 768px) {
