+++ src/components/SideBar/index.scss
@@ -43,7 +43,7 @@
 
     @media screen and (min-width: 600px) {
       height: 98%;
-      padding-bottom: 2%;
+      padding-bottom: $side-bar-padding;
       margin: initial;
     }
 
@@ -80,7 +80,8 @@
 
     @media screen and (min-width: 600px) {
       height: 98%;
-      padding-bottom: 2%;
+      margin-top: 2px;
+      padding-bottom: $side-bar-padding;
     }
 
     @media screen and (max-width: 600px) {
@@ -92,7 +93,7 @@
     height: 15px;
     width: 15px;
     border: 3px solid black;
-    margin-bottom: $base-font-size;
+    margin-bottom: $base-font-size * 1.2;
     transition-property: border-radius;
     transition-duration: .5s;
     cursor: pointer;
