+++ src/containers/PortfolioRoot/index.scss
@@ -14,7 +14,7 @@
   }
 
   @media screen and (min-width: 768px) {
-    height: 85vh;
+    height: calc(100vh - 70px);
   }
 
   &__container {
