+++ src/variables.scss
@@ -11,3 +11,5 @@ $header-height: 65px;
 $header-height__tablet: $header-height + 5px;
 $header-height-spacing: 135px;
 $header-height-spacing__tablet: ($header-height__tablet * 1.5);
+
+$side-bar-padding: calc(2% + .5vw);
